Extract toolbar logic from drawer activity
